Move tests into fitting namespaces for UI tests, utils and services.
