+++ FHICORC.Tests/FHICORC.Tests.csproj
@@ -90,8 +90,10 @@
     <Folder Include="ViewModelTests\" />
     <Folder Include="BusinessRules\" />
     <Folder Include="Factories\" />
+    <Folder Include="UITests\" />
   </ItemGroup>
   <ItemGroup>
     <None Remove="Factories\" />
+    <None Remove="UITests\" />
   </ItemGroup>
 </Project>
